+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Default response functionality
 - Default exception functionality
+- `getLastRequest` method
 
 ## 1.0.1 - 2017-05-02